+++ front/src/components/atoms/embedding/CloudOne/CloudOne.jsx
