+++ front/src/components/atoms/embedding/CloudOne/CloudOne.style.js
