+++ front/src/components/atoms/embedding/CloudSecond/CloudSecond.jsx
