+++ front/src/components/atoms/embedding/CloudSecond/CloudSecond.style.js
