+++ front/src/components/atoms/embedding/MountainIcon/MountainIcon.jsx
