+++ front/src/components/atoms/embedding/MountainIcon/MountainIcon.style.js
